feat: 쿠폰 CRD Controller 구현

+++ src/main/java/com/example/BE/coupon/service/CouponService.java
@@ -3,6 +3,8 @@ package com.example.BE.coupon.service;
 import com.example.BE.coupon.controller.dto.CouponCreateRequest;
 import com.example.BE.coupon.controller.dto.CouponResponse;
 import com.example.BE.coupon.repository.CouponRepository;
+import com.example.BE.global.exception.CustomException;
+import com.example.BE.global.exception.errorCode.CouponErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -26,6 +28,10 @@ public class CouponService {
 
     @Transactional
     public void delete(Long couponId) {
+        if (!couponRepository.existsById(couponId)) {
+            throw CustomException.from(CouponErrorCode.NOT_FOUND);
+        }
+        // TODO: userCoupon 먼저 지우기
         couponRepository.deleteById(couponId);
     }
